# Remove reflect.DeepEqual from listener hot path