--- a/pilot/pkg/networking/core/v1alpha3/listener_test.go
+++ b/pilot/pkg/networking/core/v1alpha3/listener_test.go
@@ -626,6 +626,15 @@ func TestGetActualWildcardAndLocalHost(t *testing.T) {
 	}
 }
 
+func TestIsFilterChainMatchEmpty(t *testing.T) {
+	fcm := listener.FilterChainMatch{}
+	e := reflect.ValueOf(&fcm).Elem()
+	// This isn't really testing the code, its just making sure an Envoy update won't silently break this method
+	if e.NumField() != 13 {
+		t.Fatalf("Expected 13 fields, got %v. This means we need to update isFilterChainMatchEmpty", e.NumField())
+	}
+}
+
 func testOutboundListenerConflictWithSniffingDisabled(t *testing.T, services ...*model.Service) {
 	t.Helper()