@istio-testing 2020-04-06
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
lint_istio | ad306d32ae211ff9e54c876428251493ad80adc1 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/22752/lint_istio/10890) | `/test lint_istio`
gencheck_istio | ad306d32ae211ff9e54c876428251493ad80adc1 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/22752/gencheck_istio/9469) | `/test gencheck_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->